pyproject.toml
[tool.poetry]
name = "cbus-throttle-display"
version = "0.1.0"
description = ""
authors = ["Tom Homewood"]

[tool.poetry.dependencies]
python = "^3.8"
python-can = "^4.0.0"
rich = "^12.2.0"
requests = "^2.27.1"
PySimpleGUI = "^4.59.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
